MINOR: ignore heartbeat response if leaving (#15362)
When the consumer enters state LEAVING, it sets the epoch to the leave epoch,
such as -1. When the timing is right, we may get a heartbeat response after
entering the state LEAVING, which resets the epoch to the member epoch on
the server. The result is that the consumer never leaves the group.

Seems like fd819dcee7775627341cd1e35aa125ac0b5453f0 changed the timing inside
the consumer to relatively frequently triggers this problem inside
`DescribeConsumerGroupTest`.

We fix it by ignoring any heartbeat responses when we are in state LEAVING.

Reviewers: David Jacot <djacot@confluent.io>