KAFKA-16556: SubscriptionState should not prematurely reset 'pending' partitions (#15724)
Partitions that are marked as pendingOnAssignedCallback should not be reset in resetInitializingPositions(). Pending partitions are omitted from the set returned by initializingPartitions(). As a result, the Consumer does not include them in the set of partitions for which it attempts to load committed offsets. The code used by the Consumer to reset positions (resetInitializingPositions()) should likewise skip partitions marked as pending.

Reviewers: Lucas Brutschy <lbrutschy@confluent.io>